requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
coverage==7.2.7
dnspython==2.3.0
greenlet==2.0.2
idna==3.4
numpy==1.25.0
pandas==2.0.2
psycopg2-binary==2.9.6
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-mock==1.11.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
toml==0.10.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
